Extract AudioRouteMonitoring.java from CordovaCall.java
