+++ plugin.xml
@@ -67,6 +67,10 @@
         <source-file src="src/android/MyConnectionService.java" target-dir="src/com/dmarc/cordovacall" />
         <source-file src="src/android/CallAudioService.java" target-dir="src/com/dmarc/cordovacall" />
         <source-file src="src/android/PhoneAccountManager.java" target-dir="src/com/dmarc/cordovacall" />
+        <source-file src="src/android/CallConnection.java" target-dir="src/com/dmarc/cordovacall" />
+        <source-file src="src/android/IncomingCallConnection.java" target-dir="src/com/dmarc/cordovacall" />
+        <source-file src="src/android/OutgoingCallConnection.java" target-dir="src/com/dmarc/cordovacall" />
+        <source-file src="src/android/AudioRouteMonitoring.java" target-dir="src/com/dmarc/cordovacall" />
   </platform>
 
     <platform name="ios">
